refine pretty_send and clean up imports
