+++ src/jobs/fill_posts_list_job.py
@@ -7,9 +7,10 @@ from ..app_context import AppContext
 from ..consts import TrelloListAlias, TrelloCardColor
 from ..sheets.sheets_objects import RegistryPost
 from ..strings import load
+from ..tg.sender import pretty_send
 from ..trello.trello_client import TrelloClient
 from .base_job import BaseJob
-from .utils import check_trello_card, format_errors, pretty_send
+from .utils import check_trello_card, format_errors
 
 logger = logging.getLogger(__name__)
