+++ src/jobs/publication_plans_job.py
@@ -6,9 +6,10 @@ from typing import Callable, List
 from ..app_context import AppContext
 from ..consts import TrelloListAlias, TrelloCardColor
 from ..strings import load
+from ..tg.sender import pretty_send
 from ..trello.trello_client import TrelloClient
 from .base_job import BaseJob
-from .utils import check_trello_card, format_errors, format_possibly_plural, pretty_send
+from .utils import check_trello_card, format_errors, format_possibly_plural
 
 logger = logging.getLogger(__name__)
