+++ src/jobs/tg_analytics_report_job.py
@@ -1,13 +1,12 @@
 import json
 import logging
-from datetime import datetime, timedelta, timezone
+from datetime import datetime, timezone
 from typing import Callable, Tuple
 
 from ..strings import load
-
+from ..tg.sender import pretty_send
 from src.app_context import AppContext
 from src.jobs.base_job import BaseJob
-from src.jobs.utils import pretty_send
 
 logger = logging.getLogger(__name__)
