+++ src/tg/sender.py
@@ -1,14 +1,31 @@
 """Sends messages"""
 
 import logging
+import time
 from typing import Callable, List
 
 import telegram
 
+from ..consts import MESSAGE_DELAY_SEC
 from ..utils.singleton import Singleton
 
 logger = logging.getLogger(__name__)
 
+
+def pretty_send(
+        paragraphs: List[str],
+        send: Callable[[str], None]
+) -> str:
+    '''
+    Send a bunch of paragraphs grouped into messages with adequate delays.
+    Return the whole message for testing purposes.
+    '''
+    messages = paragraphs_to_messages(paragraphs)
+    for i, message in enumerate(messages):
+        if i > 0:
+            time.sleep(MESSAGE_DELAY_SEC)
+        send(message)
+    return '\n'.join(messages)
 
 class TelegramSender(Singleton):
     def __init__(
@@ -62,13 +79,17 @@ class TelegramSender(Singleton):
         Sends a message to a single chat_id.
         """
         try:
-            self.bot.send_message(
-                text=message_text.strip(),
-                chat_id=chat_id,
-                disable_notification=self.is_silent,
-                disable_web_page_preview=self.disable_web_page_preview,
-                parse_mode=telegram.ParseMode.HTML,
-                **kwargs
+            pretty_send(
+                [message_text.strip()],
+                lambda msg:
+                    self.bot.send_message(
+                        text=msg,
+                        chat_id=chat_id,
+                        disable_notification=self.is_silent,
+                        disable_web_page_preview=self.disable_web_page_preview,
+                        parse_mode=telegram.ParseMode.HTML,
+                        **kwargs
+                    )
             )
             return True
         except telegram.TelegramError as e:
@@ -106,3 +127,47 @@ class TelegramSender(Singleton):
         )
         self.is_silent = self._tg_config.get('is_silent', True)
         self.disable_web_page_preview = self._tg_config.get('disable_web_page_preview', True)
+
+
+def paragraphs_to_messages(
+        paragraphs: List[str],
+        char_limit=telegram.constants.MAX_MESSAGE_LENGTH,
+        delimiter='\n\n',
+) -> List[str]:
+    '''
+    Makes as few message texts as possible from given paragraph list.
+    '''
+    if not paragraphs:
+        logger.warning('No paragraphs to process, exiting')
+        return []
+
+    delimiter_len = len(delimiter)
+    messages = []
+    paragraphs_in_message = []
+    char_counter = char_limit  # so that we start a new message immediately
+
+    for paragraph in paragraphs:
+        if len(paragraph) + char_counter + delimiter_len < char_limit:
+            paragraphs_in_message.append(paragraph)
+            char_counter += len(paragraph) + delimiter_len
+        else:
+            # Overflow, starting a new message
+            messages.append(delimiter.join(paragraphs_in_message))
+            # if paragraph is too long, force split it for line breaks
+            while len(paragraph) >= char_limit:
+                # last line break before limit
+                last_break_index = paragraph[:char_limit].rfind('\n')
+                if last_break_index == -1:
+                    # if no line breaks, force split
+                    last_break_index = char_limit - 1
+                messages.append(paragraph[:last_break_index])
+                paragraph = paragraph[last_break_index:].strip()
+            # start new message
+            paragraphs_in_message = [paragraph]
+            char_counter = len(paragraph)
+    # add last remaining message to the list
+    messages.append(delimiter.join(paragraphs_in_message))
+
+    # first message is empty by design.
+    assert messages[0] == ''
+    return messages[1:]
